how_to_become_football_mvp/requirements.txt
ansi2html==1.9.1
attrs==23.1.0
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
dash==2.14.2
dash-bootstrap-components==1.5.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
fiona==1.9.5
flask==3.0.0
geopandas==0.13.2
idna==3.6
importlib-metadata==7.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
nest-asyncio==1.5.8
numpy==1.24.4
packaging==23.2
pandas==2.0.3
Pillow==10.1.0
plotly==5.18.0
pyproj==3.5.0
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
retrying==1.3.4
shapely==2.0.2
six==1.16.0
tenacity==8.2.3
typing-extensions==4.9.0
tzdata==2023.3
urllib3==2.1.0
werkzeug==3.0.1
zipp==3.17.0
